backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
python-dotenv==0.19.0
google-generativeai==0.3.1
PyPDF2==3.0.1
pydantic==1.10.2
